refactor: pickle utils
